requirements.txt
asgiref==3.6.0
boto3==1.26.111
botocore==1.29.111
dj-database-url==1.3.0
Django==3.0.5
django-filter==23.1
django-storages==1.13.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
jmespath==1.0.1
mongoengine==0.27.0
Pillow==9.5.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
